arranged authenticatin files
