+++ src/app/@application/(components)/clearBtn.js
@@ -1,7 +1,7 @@
 import style from '@/style/setting.module.css'
-export let ClearBtn=({text, svg, reverse})=>{
+export let ClearBtn=({text, svg, reverse, btnFunc})=>{
     return (
-        <button className={`${style.titleButton}  ${reverse && style.reverse}`}>
+        <button onClick={()=> btnFunc()} className={`${style.titleButton}  ${reverse && style.reverse}`}>
             {text}
             <div>
                 {svg && svg}
