+++ src/app/@application/(components)/propertyInfo.js
@@ -2,11 +2,14 @@
 import style from "@/style/properties.module.css"
 export default function PropertyInfo({h1Text, buttonText, buttonFunc, arrProperties, isEdit, extraButtonText, extraButtonFunc, showSvg}) {
     return(
-        <div className={style.flex+ ' ' + style.alignCenter}>
+        <div className={style.flex2 + ' ' + style.alignCenter}>
         <h1 className={style.tableH1}>{h1Text} { arrProperties?.length > 0 && <span className={style.span}>{arrProperties?.length}</span>}</h1>
-        {showSvg && <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+        {showSvg &&
+        <div>
+        <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path d="M4.66699 6.66663L8.00033 9.99996L11.3337 6.66663" stroke="#333333" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
-        </svg>}
+        </svg>
+        </div>}
         {
             buttonText &&
             <button className={`${ isEdit ? style.label : style.button}` + ' ' + style.autoLeft} onClick={()=> {
@@ -15,7 +18,7 @@ export default function PropertyInfo({h1Text, buttonText, buttonFunc, arrPropert
         }
         {
             extraButtonText && 
-            <button className={style.label} onClick={()=> {
+            <button className={style.label + ' ' + style.mobile} onClick={()=> {
                 extraButtonFunc()
             }}>{extraButtonText}</button>
         }
