+++ src/app/@application/Expenses/page.js
@@ -1,5 +1,6 @@
 'use client'
 
+import { useRouter } from "next/navigation";
 import TopBar from "../(components)/Topbar";
 import DocumentHeader from "../(components)/documentHeader";
 import { ExpenseFooter } from "../(components)/expenseFooter";
@@ -8,6 +9,7 @@ import ExpenseTitle from "../(components)/expenseTitle";
 import PropertyInfo from "../(components)/propertyInfo";
 
 export default function Expenses(params) {
+    let router = useRouter()
     let filterSvg = <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path opacity="0.15" d="M2.66675 2.66669L6.00008 8.00002V12L10.0001 14V8.00002L13.3334 2.66669H2.66675Z" fill="#A61D4A"/>
     <path d="M2.66675 2.66669L6.00008 8.00002V12L10.0001 14V8.00002L13.3334 2.66669H2.66675Z" stroke="#A61D4A" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
@@ -17,7 +19,7 @@ export default function Expenses(params) {
     </svg>    
     return (
         <>
-        <TopBar title={`Expenses`} btnFunc={()=> console.log('called')} buttonText={"Add an expense"} showSearch={true} />
+        <TopBar title={`Expenses`} btnFunc1={()=> router.push('/Expenses/addExpense')} buttonText1={"Add an expense"} showSearch={true} />
         <PropertyInfo h1Text={`4517 Washington Ave. Manchester, Kentucky 39495`} showSvg={true}/>
         <ExpenseTitle title={`Expenses`} svg1={filterSvg} buttonText1={`Filters`} svg2={exportSvg} buttonText2={`Export as`}/>
         <ExpensesTable />
