+++ src/app/@application/Properties/[id]/page.js
@@ -2,9 +2,9 @@ import EditProperty from "../../(hoc)/editPage";
 export default async function EditPbnage(params) {
     let id = params.params.id
     return(
-        <div>
+        <>
         <EditProperty id={id}/>
-        </div>
+        </>
     )
 }
