+++ src/app/@authentication/(presentation)/divOr.js
@@ -1,4 +1,4 @@
-import style from "../../../style/auth.module.css"
+import style from "@/style/auth.module.css"
 export default function DivOr(params) {
     return(
         <div className={style.or}>
