+++ src/app/@authentication/page.js
@@ -1,4 +1,4 @@
-import SignIn from "./pages/login";
+import SignIn from "./(pages)/login";
 export const metadata = {
     title: 'Keepingly'
 }
