+++ src/style/expense.module.css
@@ -32,7 +32,7 @@ gap: 1rem;
 .left{
     flex: 1 0 0;    
 }
-.table{
+.table, .table2{
     display: flex;
     flex-direction: column;
     width: 100%;
@@ -43,7 +43,6 @@ gap: 1rem;
     justify-content: space-between;
     width: 100%;
     display: flex;
-    position: relative;
 align-items: flex-start;
 gap: 0.5rem;
 align-self: stretch;
@@ -79,6 +78,7 @@ line-height: 1.25rem; /* 142.857% */
 padding: 0.625rem 1rem;
 align-items: center;
 align-self: stretch;
+position: relative;
 background: var(--black-0, #FFF);
 }
 .small{
@@ -94,8 +94,8 @@ flex-direction: column;
 justify-content: center;
 align-items: center;
 position: absolute;
-right:  -0rem;
-top: 5rem;
+right: 0;
+z-index: 100;
 border-radius: 0.5rem;
 background: var(--black-0, #FFF);
 box-shadow: 0px 8px 16px 0px rgba(51, 51, 51, 0.08);
@@ -235,4 +235,57 @@ font-size: 1rem;
 font-style: normal;
 font-weight: 400;
 line-height: 1.5rem; /* 150% */
+}
+.display{
+    display: block;
+}
+.noDisplay{
+    display: none;
+}
+.multiForm{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+.renovationHold{
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    width: 100%;
+    height: 10%;
+}
+@media (max-width: 800px) {
+    .table{
+        overflow-x: scroll;
+        overflow-y: hidden;
+    }
+    .table2{
+        width: 1500px;
+    }
+    .footerText{
+        text-wrap: nowrap;
+    }
+    .footer{
+        gap: .2rem;
+    }
+    .span{
+        text-wrap: nowrap;
+    }
+    .noDisplay{
+        display: block;
+    }
+    .stage{
+        display: none;
+    }
+    .divider{
+        gap: 0.5rem;
+    }
+    .multiForm{
+        margin-top: 10px;
+    }
+    .header{
+        flex-direction: column;
+        align-items: flex-start;
+    }
 }
