+++ src/style/properties.module.css
@@ -40,7 +40,8 @@
     align-items: center;
     gap: 1rem;
     background-color: #FFF;
-    width: fit-content;
+    width: 100%;
+    height: 100%;
     box-sizing: border-box;
     border-radius: 5px;
     border: 1px solid #AAAAAA;
@@ -48,6 +49,7 @@
 }
 .input input, .input input::placeholder{
     font-size: 14px;
+    width: 100%;
     font-weight: 400;
     line-height: 20px;
     letter-spacing: 0em;
@@ -55,9 +57,10 @@
     text-align: left;
     color: #AAAAAA;
 }
-.header div{
+.headerdiv, .mobile{
     display: flex;
     align-items: center;
+    height: 100%;
     gap: 8px;
 }
 .header img{
@@ -65,7 +68,7 @@
     height: 40px;
     border-radius: 24px;
 }
-.flex{
+.flex, .flex2{
     align-items: center;
     display: flex;
     gap: 32px;
@@ -247,6 +250,7 @@
     position: fixed;
     top: 0;
     left: 0;
+    z-index: 1000;
     align-items: center;
     background: #838383bb;
     justify-content: center;
@@ -460,4 +464,73 @@
 }
 .noView{
     opacity: 0;
+}
+@media (max-width: 800px) {
+    .flex{
+        gap: 1.5rem;
+        flex-direction: column;
+        margin-top: 1.5rem;
+    }
+    .flex2{
+        margin-top: 1.5rem;
+    }
+    .upload{
+        width: 100%;
+        gap: 1.5rem;
+    }
+    .infoSect{
+        width: 100%;
+    }
+    .flexdiv{
+        flex-direction: column;
+    }
+    .inputLength1, .inputLength2{
+        width: 100%;
+    }
+    .modal div{
+        display: flex;
+width: 90%;
+padding: 2.5rem;
+flex-direction: column;
+align-items: center;
+gap: 1.5rem;
+    }
+    .mobile{
+        display: none;
+    }
+    .header{
+        width: 100%;
+        box-sizing: border-box;
+        height: fit-content;
+        padding-bottom: 1.5rem;
+    }
+    .headerdiv{
+        width: 100%;
+        height: 100%;
+        justify-content: space-between;
+    }
+    .width100{
+        width: 100%;
+    }
+    .tableH1{
+        width: 100%;
+        text-align: left;
+        white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    max-width: 600px;
+    }
+    .row{
+        grid-template-columns: 1fr;
+    }
+    .button{
+        width: max-content;
+        text-wrap: nowrap;
+        height: fit-content;
+    }
+    .label{
+        width: max-content;
+        text-wrap: nowrap;
+    }
+
 }
